extract knowledge base to module
