+++ client/src/components/ai-chatbot.tsx
@@ -4,6 +4,7 @@ import { Card, CardContent, CardHeader } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { MessageCircle, X, Minus, Send, Bot, User, Zap } from "lucide-react";
 import { cn } from "@/lib/utils";
+import { searchKnowledgeBase, getGenericFallback } from "@/lib/chatbot-knowledge-base";
 
 interface Message {
   id: string;
@@ -28,89 +29,6 @@ const quickActions = [
   { text: "Theme not working", icon: "🎨" },
   { text: "Mobile issues", icon: "📱" },
   { text: "Analytics help", icon: "📈" },
-];
-
-interface KnowledgeBaseEntry {
-  keywords: string[];
-  response: string;
-}
-
-const knowledgeBase: KnowledgeBaseEntry[] = [
-  {
-    keywords: ["analytics", "engagement", "metrics"],
-    response:
-      "📊 For analytics insights, visit your Analytics page where you can view engagement trends, click-through rates, and performance metrics. I can also help optimize your content strategy based on your data.",
-  },
-  {
-    keywords: ["profile", "optimize", "bio"],
-    response:
-      "✨ To optimize your profile: 1) Add a compelling bio under 160 characters, 2) Use a professional profile image, 3) Connect all your social platforms, 4) Enable pitch mode for better presentation. Need help with any specific area?",
-  },
-  {
-    keywords: ["social score", "score"],
-    response:
-      "🎯 Your Social Score is calculated based on profile completeness, link engagement, social connections, and content quality. To improve it: complete your profile, add more social links, and encourage clicks on your content.",
-  },
-  {
-    keywords: ["links", "add link", "social media"],
-    response:
-      "🔗 To manage your links: Go to 'My Links' section where you can add social media profiles, custom links, and referral links. I recommend organizing them by priority and using compelling descriptions.",
-  },
-  {
-    keywords: ["collaboration", "networking"],
-    response:
-      "🤝 Use the Collaboration section to connect with other creators, showcase spotlight projects, and manage partnership requests. This helps expand your network and create meaningful professional relationships.",
-  },
-  {
-    keywords: ["theme", "design", "appearance"],
-    response:
-      "🎨 Customize your appearance in Themes where you can choose colors, fonts, and layout styles. Pick themes that match your brand identity and professional image.",
-  },
-  {
-    keywords: ["pitch mode", "presentation"],
-    response:
-      "🎤 Pitch Mode creates a professional presentation of your profile. Configure it in Settings to highlight your expertise, focus areas, and key achievements for potential clients or collaborators.",
-  },
-  {
-    keywords: ["oauth", "connection", "connect", "disconnect"],
-    response:
-      "🔐 OAuth connections let MyLinked securely access social platforms. Visit Settings > Social Connections to connect or disconnect accounts. Ensure proper permissions are granted and check platform-specific requirements.",
-  },
-  {
-    keywords: ["api", "developer"],
-    response:
-      "🛠️ Our platform uses several APIs for social media integration. If you're experiencing API issues, verify your credentials, rate limits, and ensure the platform's API status is operational.",
-  },
-  {
-    keywords: ["privacy", "data", "security"],
-    response:
-      "🔒 We protect your data with industry-standard security including encrypted storage and secure OAuth flows. Control what you share by managing connected accounts and links in Settings.",
-  },
-  {
-    keywords: ["account", "delete", "username", "email"],
-    response:
-      "👤 Account settings allow updating username, email, or requesting account deletion via Support. Navigate to Settings to manage these options.",
-  },
-  {
-    keywords: ["problem", "issue", "not working"],
-    response:
-      "🔧 I can help troubleshoot issues! Common solutions: 1) Refresh the page, 2) Clear browser cache, 3) Check your internet connection, 4) Try a different browser. What specific problem are you experiencing?",
-  },
-  {
-    keywords: ["how to", "tutorial", "guide"],
-    response:
-      "📚 I can provide step-by-step guidance for any MyLinked feature. What would you like to learn? Popular topics: setting up profiles, optimizing for engagement, using analytics, or managing collaborations.",
-  },
-  {
-    keywords: ["hello", "hi", "hey"],
-    response:
-      "Hello! I'm your intelligent MyLinked assistant. I can help you optimize your profile, understand analytics, manage social links, troubleshoot issues, and boost your social presence. What would you like to work on today?",
-  },
-  {
-    keywords: ["help", "what can you do"],
-    response:
-      "🚀 I can assist with: Profile optimization, Analytics insights, Social Score improvement, Link management, Collaboration networking, Theme customization, Pitch Mode setup, OAuth connections, and troubleshooting. What specific area interests you?",
-  },
 ];
 
 export function AIChatbot() {
@@ -231,6 +149,8 @@ export function AIChatbot() {
   };
 
   const getAIResponse = async (userInput: string): Promise<string> => {
+    const local = searchKnowledgeBase(userInput);
+    if (local) return local;
     try {
       const response = await fetch("/api/ai-support/chat", {
         method: "POST",
@@ -251,23 +171,11 @@ export function AIChatbot() {
       return data.response || "I'm having trouble processing your request right now. Please try again.";
     } catch (error) {
       console.error("AI Response Error:", error);
-      // Fallback to intelligent local responses
-      return getIntelligentFallback(userInput);
+      return getGenericFallback(userInput);
     }
   };
 
-  const getIntelligentFallback = (userInput: string): string => {
-    const input = userInput.toLowerCase();
-    const match = knowledgeBase.find((entry) =>
-      entry.keywords.some((keyword) => input.includes(keyword))
-    );
-    if (match) {
-      return match.response;
-    }
-    return `I understand you're asking about "${userInput}". I can provide specific guidance on MyLinked features, optimization strategies, or troubleshooting. Could you tell me more about what you're trying to accomplish or what specific help you need?`;
-  };
-
-  const handleKeyPress = (e: React.KeyboardEvent) => {
+  const handleKeyDown = (e: React.KeyboardEvent) => {
     if (e.key === "Enter" && !e.shiftKey) {
       e.preventDefault();
       handleSendMessage();
@@ -441,7 +349,7 @@ export function AIChatbot() {
                 onChange={(e) =>
                   setState((prev) => ({ ...prev, inputValue: e.target.value }))
                 }
-                onKeyPress={handleKeyPress}
+                onKeyDown={handleKeyDown}
                 placeholder="Type your message..."
                 disabled={state.isLoading}
                 className="flex-1"
